Ported PF algoritm and fixed LL
The A* pathfinding algoritm is now fully implemented and some problems with LinkedList.h was fixed. For the pathfinding to work correctly an enemy needs to know the Integer values of the Tile it is on but this has nothing to to with the A* algorithm. A* seems to work as intended